Moved ExitReturn into it's own file.
