+++ core/__init__.py
@@ -1,2 +1,3 @@
 from . import logger
 from . import cli_entry
+from . import exit_return
